Fix equals and hashcode implementations for keys
The previous implementation for both Key and KeyCombination was wrong
because the resulting hash code wasn't the same when switching keys
around but the `Equals`-method returned true regardless, but according
to the official documentation the hash code must be the same for any
objects that are equal.

See also:
- https://learn.microsoft.com/en-us/dotnet/api/system.object.gethashcode?view=net-7.0#notes-to-inheritors
- https://stackoverflow.com/questions/720177/default-implementation-for-object-gethashcode
- https://thomaslevesque.com/2020/05/15/things-every-csharp-developer-should-know-1-hash-codes/

The second resource especially helped a lot with my custom hash code
implementations.